requirements.txt
python3-indy
psycopg2-binary
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.39
Flask-Cors==3.0.9
flask-jwt-extended==3.25.0
bcrypt==4.0.0
flask-bcrypt==1.0.1
markupsafe==2.0.1
PyJWT==1.6.4
